coffee.py
class Coffee:
    def __init__(self, name):
        self.name = name
        self._orders = []

    @property
    def name(self):
        return self._name

    @name.setter
    def name(self, value):
        if not isinstance(value, str) or len(value) < 3:
            raise ValueError("Name must be a string with at least 3 characters.")
        self._name = value

    def orders(self):
        return self._orders

    def customers(self):
        return list(set(order.customer for order in self._orders))

    def num_orders(self):
        return len(self._orders)

    def average_price(self):
        if self.num_orders() == 0:
            return 0
        total_price = sum(order.price for order in self._orders)
        return total_price / self.num_orders()
